doc-forgery-main/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9
torch>=2.2
segmentation-models-pytorch>=0.3.3
timm>=1.0.0
numpy>=1.26
opencv-python-headless>=4.9
Pillow>=10.3
imagehash>=4.3
PyMuPDF>=1.24
easyocr>=1.7
pytest>=8.2
